requirements.txt
Django==1.11.3
django-memoize==2.1.0
mysqlclient==1.3.10
names==0.3.0
python-memcached==1.58
pytz==2017.2
six==1.10.0
